==> 2. Sampling V2.ipynb <==
# 2. Sampling Version 2 (OMITTED)

As we've defined in the proposal, we will be studying two seperate samples. One from which we include videos from selected, and one which dives into the general domain. This section will regard the latter variant.

**EDIT:**  
This section was eventually scrapped and omitted from the actualy thesis because the videos retrieved did not concur with the sampling definition. We found it is rather difficult to accurate sample videos without passing channel ids. The code below shows how you could go about doing it; however, controlling your results will be an important challenge you will need to navigate.

**EDIT 2:**  
Also, please note this section was written at the very start of the thesis, using one singular function would make this code much more efficient. 

# Overall Document Prep:
from pyforest import * # This library quickly imports most of the relevant Data Science libraries
directory = '####'     # Set a working directory

## 2.1 Set-up the Service

We'll be using the Youtube API to select videos.

# Allocate credentials:
from googleapiclient.discovery import build

# Api Keys
key = "#####"

# Session Build
youtube = build('youtube', 'v3', developerKey = key)

## 2.2 Generate the Sample Using the Service

Below we generate the sample by using the service. We will be searching based on Keywords as well as video ID's both will retrieve videos related to either parameter.

# Variant 2 Sample Generation
keywords       = ['Global Warming',
                  'Coronavirus',
                  'Mars Rover',
                  'Fusion Energy',
                  'Genetic engineering']
rand_selec_ids = ['IJ-h0SO8EUI',
                  'LEZCxxKp0hM',
                  'Y52e551lU50',
                  '7u_cm5b1s7Y',
                  'Lx4YuW3bagE',
                  'dMdvwoldnko',
                  'gIAsT2uDGxE',
                  'H6u0VBqNBQ8',
                  'F3QpgXBtDeo',
                  'R13BD8qKeTg']
order          = 'relevance'
iter           = range(1, 150)

Raw_sample_V2  = pd.DataFrame(columns = ["Video.ID", "Title"])

### 2.2.1 Search Based on Keywords

#Iterate through Keywords
for kw in keywords:
    
    print(f'Searching on {kw}')
    
    #Iterate iter number of times to fulfill n; there is a maximum of 50 results per search.
    for i in range(1, 150):

        if i == 1:
            
            # Search Request
            request = youtube.search().list(
                part              = "snippet",
                q                 = kw,
                type              = "video",
                relevanceLanguage = 'en',
                maxResults        = 2500,
                videoCategoryId   = 28,
                order             = order,
                eventType         = "completed",
                videoDefinition   = 'high',
                videoDuration     = "medium"
            )    
                                                # Bug: It keeps copying ids and includes low quality videos.
            # Save response
            response = request.execute()

            # Unpack Respons
            rows = []
            
            for item in response['items']:

                    rows.append([item['id']['videoId'],
                                item['snippet']['title']])

            video_sample = pd.DataFrame(rows, columns = ["Video.ID", "Title"])
            print(f'{len(video_sample)} videos retrieved!')
        
        else:
            try:   
                # Search Request
                request = youtube.search().list(
                    part              = "snippet",
                    q                 = kw,
                    type              = "video",
                    relevanceLanguage = 'en',
                    maxResults        = 2500,
                    videoCategoryId   = 28,
                    order             = order,
                    eventType         = "completed",
                    videoDuration     = "medium",
                    videoDefinition   = 'high',
                    pageToken         = response['nextPageToken']
                ) 

                # Save response
                response = request.execute()

                # Unpack Respons
                rows = []

                for item in response['items']:

                    rows.append([item['id']['videoId'],
                                item['snippet']['title']])

                video_sample_temp = pd.DataFrame(rows, columns = ["Video.ID", "Title"])
                video_sample = video_sample.append(video_sample_temp)
                print(f'{len(video_sample)} videos retrieved!')
            
            except(KeyError):
                print("\n Results exhausted \n")
                break
    
    #Cleaning:
    to_delete = ['#short', 
                 ' prank',
                 '[live]',
                 '[LIVE]',
                 'live ',
                 'news',
                 'summit']
    #video_sample = video_sample[~video_sample['Title'].str.contains('|'.join(to_delete))] 
    # Exclusion omitted due to mixed results. 
    
    #Output:
    Raw_sample_V2 = Raw_sample_V2.append(video_sample)
    print(f'The sample searched on {kw} has been saved! \n')

Raw_sample_V2_kw.to_csv(f'{directory}Raw_Sample_kw.csv', sep = ';')

### 2.3.2 Search Based on Related Video IDs

#Iterate through Related Video IDs
for rvi in rand_selec_ids:
    
    print(f'Searching {rand_selec_ids.index(rvi) +1} out of {len(rand_selec_ids)} videos')
    
    #Iterate iter number of times to fulfill n; there is a maximum of 50 results per search.
    for i in range(1, 150):

        if i == 1:
            
            # Search Request
            request = youtube.search().list(
                part              = "snippet",
                q                 = kw,
                type              = "video",
                relevanceLanguage = 'en',
                maxResults        = 2500,
                videoCategoryId   = 28,
                order             = order,
                eventType         = "completed",
                videoDefinition   = 'high',
                videoDuration     = "medium"
            )    
                                                        # It keeps copying id's and includes low quality videos.
            # Save response
            response = request.execute()

            # Unpack Respons
            rows = []
            
            for item in response['items']:

                    rows.append([item['id']['videoId'],
                                item['snippet']['title']])

            video_sample = pd.DataFrame(rows, columns = ["Video.ID", "Title"])
            print(f'{len(video_sample)} videos retrieved!')
        
        else:
            try:   
                # Search Request
                request = youtube.search().list(
                    part              = "snippet",
                    q                 = kw,
                    type              = "video",
                    relevanceLanguage = 'en',
                    maxResults        = 2500,
                    videoCategoryId   = 28,
                    order             = order,
                    eventType         = "completed",
                    videoDuration     = "medium",
                    videoDefinition   = 'high',
                    pageToken         = response['nextPageToken']
                ) 

                # Save response
                response = request.execute()

                # Unpack Respons
                rows = []

                for item in response['items']:

                    rows.append([item['id']['videoId'],
                                item['snippet']['title']])

                video_sample_temp = pd.DataFrame(rows, columns = ["Video.ID", "Title"])
                video_sample = video_sample.append(video_sample_temp)
                print(f'{len(video_sample)} videos retrieved!')
            
            except(KeyError):
                print("\n Results exhausted \n")
                break
    
    #Cleaning:
    to_delete = ['#short', 
                 ' prank',
                 '[live]',
                 '[LIVE]',
                 'live ',
                 'news',
                 'summit']
    #video_sample = video_sample[~video_sample['Title'].str.contains('|'.join(to_delete))] 
    
    #Output:
    Raw_sample_V2_rvi = Raw_sample_V2.append(video_sample)
    print(f'The sample searched on {kw} has been saved! \n')

Raw_sample_V2_rvi.to_csv(f'{directory}Raw_Sample_rvi.csv', sep = ';')
