==> 4. Transcription.ipynb <==
# 4. Transcription

There will be several parts to this section. First, we'll have to download the vidoes. Then we'll setup the API, then we'll run the transcription, then we'll decode and store the transcriptions as .txt files

# Overall Document Prep:
from pyforest import * # This library quickly imports most of the relevant Data Science libraries
directory = '####'     # Set a working directory

### 4.1 Sampling video ID's

import jupyternotify # Since these cells will take a long time to run Jupyter Notify is 
ip = get_ipython()   # helpful to make sure we are notified as soon as the processes finish.
ip.register_magics(jupyternotify.JupyterNotifyMagics)

#Load in sample
Sample_V1 = pd.read_csv(f'{directory}Sample_V1.csv', sep = ";")
#Sample_V2 = pd.read_csv(f'{directory}Sample_V2.csv', sep = ";")

## 4.2 Downloading Youtube Audio

# To download the audio files from youtube videos we'll be using the Youtube-dl library
# This library works on the command line, however we can use subprocess to control a shell
# at the command line with python through Jupyter notebooks. 
import subprocess
import glob # To save time, we try to exclude videos we already downloaded.
downloaded = glob.glob(f"/Volumes/Samsung_T5/Thesis/Audio/{variant}/*.webm")
downloaded = [item.replace(f"/Volumes/Samsung_T5/Thesis/Audio/{variant}/", "") for item in downloaded]
downloaded = [item.replace(".webm", "") for item in downloaded]

to_dl = Sample_V1[~Sample_V1['Video.ID'].isin(downloaded)]
to_dl.reset_index(drop=True,inplace=True)

for i in to_dl['Video.ID']:
    command = f'youtube-dl -f 251 http://www.youtube.com/watch?v={i} -o /Volumes/Samsung_T5/Thesis/Audio/V1/{i}.webm'
    subprocess.call(command, shell = True)
    
    # Progress report
    current = to_dl[to_dl['Video.ID'] == i].index[0] + 1
    every   = len(to_dl['Video.ID'])
    
    print(f'{current} out of {every} audio files has been downloaded.')
    print(f'We are {((current/every)*100):.1f}% of the way there! \n')

## 4.3 Transcribing the Youtube Audio

Adapted from: https://github.com/nicknochnack/VideoToText/blob/master/Video%20to%20Text.ipynb

### 4.3.1 Set-up Service

from ibm_watson import SpeechToTextV1
from ibm_watson.websocket import RecognizeCallback, AudioSource
from ibm_cloud_sdk_core.authenticators import IAMAuthenticator
# Setup credentials
key = '#########'
url = 'https://api.eu-gb.speech-to-text.watson.cloud.ibm.com/instances/#########'

# Setup service
authenticator = IAMAuthenticator(key)
stt = SpeechToTextV1(authenticator = authenticator)
stt.set_service_url(url)

### 4.3.2 Define the Transcribe() Function to Create .txt Files with Transcriptions

def Transcribe(df, variant='V1'):
    
    # First up, let's make sure that all of the Video ID's we pass have actually also
    # been succesfully downloaded; this is in an effort to reduce errors. 
    import glob
    downloaded = glob.glob(f"/Volumes/Samsung_T5/Thesis/Audio/{variant}/*.webm")
    downloaded = [item.replace(f"/Volumes/Samsung_T5/Thesis/Audio/{variant}/", "") for item in downloaded]
    downloaded = [item.replace(".webm", "") for item in downloaded]
    
    df = df[df['Video.ID'].isin(downloaded)]
    
    # Since the API limits the number of calls we can give, it is important to conserve
    # our calling capacity. To do so, we make sure to only include videos that are yet
    # to be transcribed. We include this code in our loop so it redfines at every iteration
    # to help us track progress.
    transcribed = glob.glob(f"/Volumes/Samsung_T5/Thesis/Text/{variant}/*.txt")
    transcribed = [item.replace(f"/Volumes/Samsung_T5/Thesis/Text/{variant}/", "") for item in transcribed]
    transcribed = [item.replace(".txt", "") for item in transcribed]
    
    df = df[~df['Video.ID'].isin(transcribed)]
    
    # Second, Let's get transcribing!
    
    # Initialize an empty dictionary to store results:
    res = {}

    for i in df['Video.ID']:
        
        print(i)
        
        with open(f'/Volumes/Samsung_T5/Thesis/Audio/{variant}/{i}.webm', 'rb') as f:
            res[f'{i}'] = stt.recognize(audio=f, 
                                        content_type='audio/webm', 
                                        model='en-US_NarrowbandModel', 
                                        continuous=True).get_result()
    
    # Third, we ought to process the results into .txt documents
        try:
            results = res[f'{i}']['results']

            text = [result['alternatives'][0]['transcript'].rstrip() + '.\n' for result in res[i]['results']]
            text = [para[0].title() + para[1:] for para in text]

            transcript = ''.join(text)

            with open(f'/Volumes/Samsung_T5/Thesis/Text/V1/{i}.txt', 'w') as out:
                out.writelines(transcript)

        except(KeyError):
            pass
    
        # Progress report
        transcribed = glob.glob(f"/Volumes/Samsung_T5/Thesis/Text/{variant}/*.txt")
        transcribed = [item.replace(f"/Volumes/Samsung_T5/Thesis/Text/{variant}/", "") for item in transcribed]
        transcribed = [item.replace(".txt", "") for item in transcribed]
        
        current = len(transcribed) + 1
        every   = len(downloaded)

        print(f'{current} out of {every} files has been transcribed.')
        print(f'We are {((current/every)*100):.1f}% of the way there! \n')
    
    return

### 4.3.3 Carrying Out the Transcription

%%notify

Transcribe(Sample_V1)
